move files around first
