+++ lifelines/estimation.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-from lifelines._base_fitter import BaseFitter
-from lifelines.weibull_fitter import WeibullFitter
-from lifelines.exponential_fitter import ExponentialFitter
-from lifelines.nelson_aalen_fitter import NelsonAalenFitter
-from lifelines.kaplan_meier_fitter import KaplanMeierFitter
-from lifelines.breslow_fleming_harrington_fitter import BreslowFlemingHarringtonFitter
-from lifelines.coxph_fitter import CoxPHFitter
-from lifelines.aalen_additive_fitter import AalenAdditiveFitter
+from lifelines.fitters import BaseFitter
+from lifelines.fitters.weibull_fitter import WeibullFitter
+from lifelines.fitters.exponential_fitter import ExponentialFitter
+from lifelines.fitters.nelson_aalen_fitter import NelsonAalenFitter
+from lifelines.fitters.kaplan_meier_fitter import KaplanMeierFitter
+from lifelines.fitters.breslow_fleming_harrington_fitter import BreslowFlemingHarringtonFitter
+from lifelines.fitters.coxph_fitter import CoxPHFitter
+from lifelines.fitters.aalen_additive_fitter import AalenAdditiveFitter
