+++ lifelines/fitters/coxph_fitter.py
@@ -8,7 +8,7 @@ from numpy.linalg import solve, norm, inv
 from scipy.integrate import trapz
 import scipy.stats as stats
 
-from lifelines._base_fitter import BaseFitter
+from lifelines.fitters import BaseFitter
 from lifelines.utils import survival_table_from_events, inv_normal_cdf, normalize,\
     significance_code, concordance_index, _get_index, qth_survival_times
