+++ lifelines/fitters/weibull_fitter.py
@@ -2,8 +2,9 @@
 from __future__ import print_function, division
 import numpy as np
 import pandas as pd
+
 from numpy.linalg import solve, norm, inv
-from lifelines._base_fitter import UnivariateFitter
+from lifelines.fitters import UnivariateFitter
 from lifelines.utils import inv_normal_cdf
 
 def _negative_log_likelihood(lambda_rho, T, E):
